API routes and react app routes protected
Users can only see the home page if and only if they are logged in to the app
- But they can see all information on the database if they log in
- Requests are not assigned to particular users

+++ frontend/src/components/RequestDetails.jsx
@@ -1,4 +1,5 @@
 import { useRequestContext } from '../hooks/useRequestsContext'
+import { useAuthContext } from "../hooks/useAuthContext";
 
 //date ffns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
@@ -6,9 +7,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const RequestDetails = ({ request }) => {
     const { dispatch } = useRequestContext();
 
+    const { user } = useAuthContext();
+
     const handleClick = async () => {
+        if(!user) {
+            return
+        }
+
         const response = await fetch('/api/requests/' + request._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json();
